animaciones y ataques

+++ IVJueguito/Assets/Game/Scripts/Enemies/IEnemy.cs
@@ -6,8 +6,8 @@ public interface IEnemy
     public void ChangeState(EnemyState newState);
     public bool IsAlive();
     public void Spawn(Vector3 pos);
-    public void TakeDamage(int damage);
-    public void DamageTarget(int damageDealt);
+    public void TakeDamage(float damage);
+    public void DamageTarget(float damageDealt);
     public void MoveTo(Vector3 target);
     public void StopMoving();
     public Vector3 SearchPlayer();
